requirements.txt
aiohttp==3.8.3
boto3==1.26.29
fsspec==2022.11.0
s3fs==0.4.2
numcodecs==0.10.2
blosc==1.10.6
tqdm==4.63.0
zarr==2.12.0
elasticdeform==0.5.0
scipy==1.7.3
torchvision==0.13.0
scikit-image==0.19.3
scikit-learn==1.0.2
tensorboard==2.11.0
ome-types==0.3.2
pandas==1.3.5
dask==2022.2.0
dask-image==2022.9.0
pytorch-msssim
debugpy
jupyterlab